Move delete_invite and add .pyi file for InviteRequest
